Began cleaning up the delayed arithmetic operations. (#1)
This introduces a UnaryIsometricOpWithArgs class that is analogous to the
DelayedUnaryIsoOpWithArgs function from the original R package. It is
currently limited to common arithmetic operations.

We split out some of the SparseNdarray-related utilities for greater re-use
across tests. SparseNdarrays are now simulated as containing NumPy arrays (this
will probably become a requirement at some point). We also added some utilities
for comparing SparseNdarrays easily.

We promoted the sparsity check to a generic, given that we aren't able
to provide a .sparse() method to NumPy arrays.