Enforce the data type of the delayed unary operation.
Users can now specify the output type of the operation (default float64) and we
always coerce the result of the operation to that type. We don't do any fancy
stuff to guess the most narrow type, to avoid overflow or precision issues.

We also avoid in-place operations if the seed type is not the same as the
specified output type, which would otherwise cause errors.