Gave up trying to protect users from NumPy's casting rules.
Now we just follow those rules and pass on the problems. So, if a user supplies
an integer seed and an value of the same type, any addition will potentially
overflow as there is no longer any casting to a larger type that can hold the
result. This simplifies our implementation and is more consistent; downstream
code has to deal with these casting rules anyway if they're using NumPy arrays.