Greatly simplified the code by switching to a dask backend. (#11)
The idea is use dask to perform all Python-side operations, thus relieving us
of the burden of implementing the rest of the NumPy array interface. The
DelayedArray wrapper classes themselves only exist to simplify the developer
experience in downstream packages like mattress and chihaya... which is all I
wanted in the first place, I didn't care about the Python side anyway.