The Subset constructor only supports sequences for simplicity.
We move slice conversion to DelayedArray.__getitem__. This means that
the subset property behaves like the argument passed in the constructor,
in that it only contains sequences of indices.

Similarly, we move the conversion of ranges to lists into the
as_dask_array function, to avoid evaluation until it is needed.