Added methods to apply operations as natively as possible. (#29)
This introduces an extract_array() function and the associate methods for each
operation. These mimic the extract_*_array() generics from the Bioconductor
package in that a subset of interest is specified for each method, and only
that subset needs to be realized. Unlike the generics, though, we also attempt
to preserve the class of the seed array, e.g., to avoid breaking sparsity.
This requires some try/catching to handle situations where the underlying
class does not support the operation or the full NumPy interface.

Added tests with sparse matrices, which were the motivation for this.