Switch to extract_sparse_array and extract_dense_array generics. (#50)
This provides more guarantees for the callers, i.e., extract_dense_array is
guaranteed to return a dense array while extract_sparse_array is guaranteed to
return a SparseNdarray. There is no longer any need to play whackamole with
bugs in various implementations (e.g., scipy.sparse); all of that pain is
concentrated in the extract_* methods for those classes, allowing the callers -
be they DelayedOp classes or end users - to operate on guaranteed objects.